Abstract remapping processing, parse antigen values outside parens
