+++ src/main/java/org/pankratzlab/unet/model/ValidationModelBuilder.java
@@ -21,8 +21,6 @@
  */
 package org.pankratzlab.unet.model;
 
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Field;
 import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -41,14 +39,10 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.TreeSet;
-import java.util.function.Function;
-import java.util.function.Predicate;
-import java.util.function.Supplier;
 import java.util.stream.Collectors;
 import javax.swing.JFrame;
 import javax.swing.JOptionPane;
 import org.apache.commons.lang3.tuple.Pair;
-import org.pankratzlab.BackgroundDataProcessor;
 import org.pankratzlab.unet.deprecated.hla.HLALocus;
 import org.pankratzlab.unet.deprecated.hla.HLAType;
 import org.pankratzlab.unet.deprecated.hla.NullType;
@@ -60,42 +54,25 @@ import org.pankratzlab.unet.hapstats.CommonWellDocumented.Status;
 import org.pankratzlab.unet.hapstats.Haplotype;
 import org.pankratzlab.unet.hapstats.HaplotypeFrequencies;
 import org.pankratzlab.unet.hapstats.RaceGroup;
-import org.pankratzlab.unet.jfx.StyleableContingentChoiceDialog;
-import org.pankratzlab.unet.jfx.StyleableContingentChoiceDialog.Option;
+import org.pankratzlab.unet.model.remap.CancellationException;
+import org.pankratzlab.unet.model.remap.RemapProcessor;
 import org.pankratzlab.unet.parser.util.BwSerotypes;
 import org.pankratzlab.unet.parser.util.BwSerotypes.BwGroup;
 import org.pankratzlab.unet.parser.util.DRAssociations;
 import org.pankratzlab.unet.parser.util.RabinKarp;
-import com.google.common.base.Predicates;
 import com.google.common.base.Strings;
 import com.google.common.collect.ArrayListMultimap;
-import com.google.common.collect.BiMap;
 import com.google.common.collect.HashBasedTable;
-import com.google.common.collect.HashBiMap;
 import com.google.common.collect.HashMultimap;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMultimap;
 import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.ImmutableSortedSet;
-import com.google.common.collect.LinkedListMultimap;
 import com.google.common.collect.ListMultimap;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Multimap;
 import com.google.common.collect.MultimapBuilder;
-import com.google.common.collect.MultimapBuilder.SortedSetMultimapBuilder;
 import com.google.common.collect.Sets;
 import com.google.common.collect.Table;
-import javafx.geometry.Pos;
-import javafx.scene.control.Label;
-import javafx.scene.control.ListCell;
-import javafx.scene.control.Tooltip;
-import javafx.scene.layout.HBox;
-import javafx.scene.layout.Priority;
-import javafx.scene.layout.Region;
-import javafx.scene.text.Text;
-import javafx.scene.text.TextFlow;
-import javafx.util.Duration;
-import javafx.util.StringConverter;
 
 /**
  * Mutable builder class for creating a {@link ValidationModel}.
@@ -131,12 +108,13 @@ public class ValidationModelBuilder {
   private Set<SeroType> aLocusCWD;
   private Set<SeroType> bLocusCWD;
   private Set<SeroType> cLocusCWD;
-  private Set<HLAType> aLocusCWDTypes;
-  private Set<HLAType> bLocusCWDTypes;
-  private Set<HLAType> cLocusCWDTypes;
   private Set<SeroType> aLocusFirst = new LinkedHashSet<>();
   private Set<SeroType> bLocusFirst = new LinkedHashSet<>();
   private Set<SeroType> cLocusFirst = new LinkedHashSet<>();
+
+  private Set<HLAType> aLocusCWDTypes;
+  private Set<HLAType> bLocusCWDTypes;
+  private Set<HLAType> cLocusCWDTypes;
   private Set<HLAType> aLocusFirstTypes;
   private Set<HLAType> bLocusFirstTypes;
   private Set<HLAType> cLocusFirstTypes;
@@ -148,25 +126,27 @@ public class ValidationModelBuilder {
   private Map<HLALocus, Pair<Set<TypePair>, Set<TypePair>>> remapping = new HashMap<>();
 
   private Set<SeroType> drbLocus;
-  private Set<SeroType> dqaLocus;
-  private Set<SeroType> dqbLocus;
-  private Set<SeroType> dpaLocus;
-
   // private Set<HLAType> drbLocusTypes;
-  private Set<HLAType> dqaLocusTypes;
-  private Set<HLAType> dqbLocusTypes;
-  private Set<HLAType> dpaLocusTypes;
-  private Set<HLAType> dpbLocusTypes;
-
-  private Set<HLAType> dqaLocusTypesNonCWD;
-  private Set<HLAType> dqbLocusTypesNonCWD;
-  private Set<HLAType> dpaLocusTypesNonCWD;
-  private Set<HLAType> dpbLocusTypesNonCWD;
-
   private Set<SeroType> drbLocusNonCWD;
+
+  private Set<SeroType> dqaLocus;
+  private Set<HLAType> dqaLocusTypes;
+  private Set<HLAType> dqaLocusTypesNonCWD;
   private Set<SeroType> dqaLocusNonCWD;
+
+  private Set<SeroType> dqbLocus;
+  private Set<HLAType> dqbLocusTypes;
+  private Set<HLAType> dqbLocusTypesNonCWD;
   private Set<SeroType> dqbLocusNonCWD;
+
+  private Set<SeroType> dpaLocus;
+  private Set<HLAType> dpaLocusTypes;
+  private Set<HLAType> dpaLocusTypesNonCWD;
   private Set<SeroType> dpaLocusNonCWD;
+
+  private Set<SeroType> dpbLocus;
+  private Set<HLAType> dpbLocusTypes;
+  private Set<HLAType> dpbLocusTypesNonCWD;
   private Set<HLAType> dpbLocusNonCWD;
 
   private Boolean bw4;
@@ -383,20 +363,21 @@ public class ValidationModelBuilder {
 
   public ValidationModelBuilder dpbNonCWD(String dpbType) {
     dpbLocusNonCWD = makeIfNull(dpbLocusNonCWD);
-    // Shorten the allele designation to allele group and specific HLA protein. Further fields can
-    // not be entered into UNOS
-    if (!Strings.isNullOrEmpty(dpbType) && dpbType.matches(".*\\d.*")) {
-      HLAType tmpDPB1 = new HLAType(HLALocus.DPB1, dpbType);
-      if (tmpDPB1.spec().size() > 2) {
-        tmpDPB1 =
-            new HLAType(HLALocus.DPB1, new int[] {tmpDPB1.spec().get(0), tmpDPB1.spec().get(1)});
-      }
-      dpbLocusNonCWD.add(tmpDPB1);
-    }
+    // // Shorten the allele designation to allele group and specific HLA protein. Further fields
+    // can
+    // // not be entered into UNOS
+    // if (!Strings.isNullOrEmpty(dpbType) && dpbType.matches(".*\\d.*")) {
+    HLAType tmpDPB1 = new HLAType(HLALocus.DPB1, dpbType);
+    // if (tmpDPB1.spec().size() > 2) {
+    // tmpDPB1 =
+    // new HLAType(HLALocus.DPB1, new int[] {tmpDPB1.spec().get(0), tmpDPB1.spec().get(1)});
+    // }
+    dpbLocusNonCWD.add(tmpDPB1);
+    // }
     return this;
   }
 
-  public ValidationModelBuilder dqb(String dqbType) {
+  public ValidationModelBuilder dqbSerotype(String dqbType) {
     if (test2(dqbType)) {
       return null;
     }
@@ -405,7 +386,7 @@ public class ValidationModelBuilder {
     return this;
   }
 
-  public ValidationModelBuilder dqa(String dqaType) {
+  public ValidationModelBuilder dqaSerotype(String dqaType) {
     if (test2(dqaType)) {
       return null;
     }
@@ -414,7 +395,7 @@ public class ValidationModelBuilder {
     return this;
   }
 
-  public ValidationModelBuilder dpa(String dpaType) {
+  public ValidationModelBuilder dpaSerotype(String dpaType) {
     if (test2(dpaType)) {
       return null;
     }
@@ -423,18 +404,23 @@ public class ValidationModelBuilder {
     return this;
   }
 
-  public ValidationModelBuilder dpb(String dpbType) {
+  public ValidationModelBuilder dpb(HLAType dpbType) {
     dpbLocusTypes = makeIfNull(dpbLocusTypes);
-    // Shorten the allele designation to allele group and specific HLA protein. Further fields can
-    // not be entered into UNOS
-    if (!Strings.isNullOrEmpty(dpbType) && dpbType.matches(".*\\d.*")) {
-      HLAType tmpDPB1 = new HLAType(HLALocus.DPB1, dpbType);
-      if (tmpDPB1.spec().size() > 2) {
-        tmpDPB1 =
-            new HLAType(HLALocus.DPB1, new int[] {tmpDPB1.spec().get(0), tmpDPB1.spec().get(1)});
-      }
-      dpbLocusTypes.add(tmpDPB1);
-    }
+    dpbLocusTypes.add(dpbType);
+    return this;
+  }
+
+  public ValidationModelBuilder dpbSerotype(String dpbType) {
+    dpbLocus = makeIfNull(dpbLocus);
+    HLAType tmpDPB1 = new HLAType(HLALocus.DPB1, dpbType);
+    dpbLocus.add(trim(tmpDPB1));
+
+    return this;
+  }
+
+  public ValidationModelBuilder dpbNonCIWD(HLAType dpbType) {
+    dpbLocusTypesNonCWD = makeIfNull(dpbLocusTypesNonCWD);
+    dpbLocusTypesNonCWD.add(dpbType);
     return this;
   }
 
@@ -564,25 +550,6 @@ public class ValidationModelBuilder {
     nonCWDLoci.add(locus.locus());
   }
 
-  private final class AlleleStringConverter extends StringConverter<Supplier<TextFlow>> {
-    private final PresentableAlleleChoices choices;
-
-    private AlleleStringConverter(PresentableAlleleChoices choices) {
-      this.choices = choices;
-    }
-
-    @Override
-    public String toString(Supplier<TextFlow> object) {
-      final String selectedData = choices.getData(object);
-      return selectedData == null ? "" : selectedData;
-    }
-
-    @Override
-    public Supplier<TextFlow> fromString(String string) {
-      return choices.getPresentation(string);
-    }
-  }
-
   public class ValidationResult {
 
     public final boolean valid;
@@ -609,9 +576,32 @@ public class ValidationModelBuilder {
     frequencyTable.clear();
 
     ValidationModel validationModel = new ValidationModel(donorId, source, sourceType, aLocusCWD,
-        bLocusCWD, cLocusCWD, drbLocus, dqbLocus, dqaLocus, dpaLocus, dpbLocusTypes, bw4, bw6,
+        bLocusCWD, cLocusCWD, drbLocus, dqbLocus, dqaLocus, dpaLocus, getFinalDPBTypes(), bw4, bw6,
         dr51Locus, dr52Locus, dr53Locus, bcCwdHaplotypes, drDqDR345Haplotypes, remapping);
     return validationModel;
+  }
+
+  // Shorten the allele designation to allele group and specific HLA protein. Further fields can
+  // not be entered into UNOS
+  private Set<HLAType> getFinalDPBTypes() {
+    Set<HLAType> dpbTypes = new HashSet<>();
+    Set<HLAType> dpbSource = dpbLocusTypes;
+    if (dpbSource == null || dpbSource.isEmpty()) {
+      dpbSource = dpbLocus.stream().map(s -> new HLAType(HLALocus.DPB1, s.spec()))
+          .collect(ImmutableSet.toImmutableSet());
+    }
+    for (HLAType dpbType1 : dpbSource) {
+      String dpbType = dpbType1.specString();
+      if (!Strings.isNullOrEmpty(dpbType) && dpbType.matches(".*\\d.*")) {
+        HLAType tmpDPB1 = new HLAType(HLALocus.DPB1, dpbType);
+        if (tmpDPB1.spec().size() > 2) {
+          tmpDPB1 =
+              new HLAType(HLALocus.DPB1, new int[] {tmpDPB1.spec().get(0), tmpDPB1.spec().get(1)});
+        }
+        dpbTypes.add(tmpDPB1);
+      }
+    }
+    return dpbTypes;
   }
 
   /**
@@ -948,14 +938,14 @@ public class ValidationModelBuilder {
   private ValidationResult ensureValidity() {
     // Ensure all fields have been set
     for (Object o : Lists.newArrayList(donorId, source, aLocusCWD, bLocusCWD, cLocusCWD, drbLocus,
-        dqbLocus, dqaLocus, dpbLocusTypes, bw4, bw6)) {
+        dqbLocus, dqaLocus, getFinalDPBTypes(), bw4, bw6)) {
       if (Objects.isNull(o)) {
         return new ValidationResult(false, Optional.of("ValidationModel incomplete"));
       }
     }
     // Ensure all sets have a reasonable number of entries
     for (Set<?> set : ImmutableList.of(aLocusCWD, bLocusCWD, cLocusCWD, drbLocus, dqbLocus,
-        dqaLocus, dpbLocusTypes)) {
+        dqaLocus, getFinalDPBTypes())) {
       if (set.isEmpty() || set.size() > 2) {
         return new ValidationResult(false,
             Optional.of("ValidationModel contains invalid allele count: " + set));
@@ -978,26 +968,11 @@ public class ValidationModelBuilder {
     return new ValidationResult(true, Optional.empty());
   }
 
-  BackgroundDataProcessor<HLALocus, PresentableAlleleChoices> choiceSupplier;
-
   public boolean hasCorrections() {
     return !nonCWDLoci.isEmpty();
   }
 
-  public boolean hasCorrectionsIfSoStartLoading() {
-    if (hasCorrections()) {
-      // start loading data in the background
-      choiceSupplier = new BackgroundDataProcessor<>(nonCWDLoci,
-          (locus) -> PresentableAlleleChoices.create(locus, getAllelePairs(locus)), (t) -> {
-            t.printStackTrace();
-            return null;
-          });
-      return true;
-    }
-    return false;
-  }
-
-  public ValidationResult processCorrections() {
+  public ValidationResult processCorrections(RemapProcessor remapProcessor) {
 
     boolean cancelled = false;
 
@@ -1005,9 +980,11 @@ public class ValidationModelBuilder {
       Pair<Set<TypePair>, Set<TypePair>> remapPair = null;
 
       try {
-        remapPair = processRemap(locus, this);
+        remapPair = remapProcessor.processRemapping(locus, this);
       } catch (CancellationException e) {
         cancelled = true;
+      } catch (Throwable t) {
+        t.printStackTrace();
       }
 
       if (remapPair != null) {
@@ -1020,21 +997,6 @@ public class ValidationModelBuilder {
     }
 
     return new ValidationResult(true, Optional.empty());
-  }
-
-  private class CancellationException extends RuntimeException {
-
-    private static final long serialVersionUID = 1L;
-
-    // used to indicate user-cancellation of locus remapping
-
-  }
-
-  interface RemapProcessor {
-
-    public Pair<Set<TypePair>, Set<TypePair>> processRemapping(HLALocus locus,
-        ValidationModelBuilder builder) throws CancellationException;
-
   }
 
   public Set<SeroType> getCWDSeroTypesForLocus(HLALocus locus) {
@@ -1051,9 +1013,14 @@ public class ValidationModelBuilder {
       case DQA1:
         return dqaLocus;
       case DPB1:
-        return null;
+        return dpbLocus;
       case DQB1:
         return dqbLocus;
+      case DRB1:
+      case DRB3:
+      case DRB4:
+      case DRB5:
+        return drbLocus;
       default:
         return null;
     }
@@ -1074,14 +1041,36 @@ public class ValidationModelBuilder {
       case DQA1:
         return dqaLocusNonCWD;
       case DPB1:
-        return dpbLocusNonCWD.stream().map(HLAType::equivSafe)
-            .collect(ImmutableSet.toImmutableSet());
+        return trim(dpbLocusNonCWD);
       case DQB1:
         return dqbLocusNonCWD;
+      case DRB1:
+      case DRB3:
+      case DRB4:
+      case DRB5:
+        return drbLocusNonCWD;
       default:
         return null;
     }
 
+  }
+
+  private Set<SeroType> trim(Set<HLAType> set) {
+    return set.stream().map(this::trim).collect(Collectors.toCollection(HashSet::new));
+  }
+
+  private SeroType trim(HLAType hlaType) {
+    List<String> seroStr = new ArrayList<>();
+    seroStr.add(String.valueOf(hlaType.spec().get(0)));
+
+    if (Objects.equals(HLALocus.DPA1, hlaType.locus())
+        || Objects.equals(HLALocus.DPB1, hlaType.locus())) {
+      // DPA and DPB report two fields
+      seroStr.add(String.valueOf(hlaType.spec().get(1)));
+    }
+
+    String[] array = seroStr.toArray(new String[seroStr.size()]);
+    return new SeroType(hlaType.locus().sero(), array);
   }
 
   public Set<HLAType> getCWDTypesForLocus(HLALocus locus) {
@@ -1101,6 +1090,10 @@ public class ValidationModelBuilder {
         return dpbLocusTypes;
       case DQB1:
         return dqbLocusTypes;
+      case DRB1:
+      case DRB3:
+      case DRB4:
+      case DRB5:
       default:
         return null;
     }
@@ -1123,625 +1116,25 @@ public class ValidationModelBuilder {
         return dpbLocusTypesNonCWD;
       case DQB1:
         return dqbLocusTypesNonCWD;
+      case DRB1:
+      case DRB3:
+      case DRB4:
+      case DRB5:
       default:
         return null;
     }
   }
 
-  private AllelePairings getPossibleAllelePairsForLocus(HLALocus locus) {
+  public AllelePairings getPossibleAllelePairsForLocus(HLALocus locus) {
     return possibleAllelePairings.get(locus);
   }
 
-  private AllelePairings getDonorAllelePairsForLocus(HLALocus locus) {
+  public AllelePairings getDonorAllelePairsForLocus(HLALocus locus) {
     return donorAllelePairings.get(locus);
   }
 
-  private static class GUIRemapProcessor implements RemapProcessor {
-
-    private Pair<Set<TypePair>, Set<TypePair>> processRemap(HLALocus locus,
-        ValidationModelBuilder builder) {
-      AllelePairings allelePairs = builder.getPossibleAllelePairsForLocus(locus);
-      AllelePairings donorPairs = builder.getDonorAllelePairsForLocus(locus);
-
-      Set<SeroType> locusSet = builder.getCWDSeroTypesForLocus(locus);
-      Set<SeroType> locusSetNonCWD = builder.getAllSeroTypesForLocus(locus);
-      Set<HLAType> typesSet = builder.getCWDTypesForLocus(locus);
-      Set<HLAType> typesSetNonCWD = builder.getAllTypesForLocus(locus);
-
-      if (allelePairs == null || donorPairs == null) {
-        // TODO
-        return null;
-      }
-
-      // FIXME TODO this needs to be done on a Task thread to show a progress spinner if it blocks
-      // while loading. However, the choice dialogs need to be created on the JavaFX thread.
-      // But also, this method overall needs to return a single validation result object ...
-      PresentableAlleleChoices choices = choiceSupplier.get(locus);
-
-      Iterator<HLAType> typeIter;
-      Iterator<SeroType> seroIter;
-
-      typeIter = typesSet.iterator();
-      HLAType hlaType1_CIWD = typeIter.next();
-      HLAType hlaType2_CIWD = typeIter.hasNext() ? typeIter.next() : hlaType1_CIWD;
-
-      // seroIter = locusSet.iterator();
-      // SeroType seroType1_CIWD = seroIter.next();
-      // SeroType seroType2_CIWD = seroIter.hasNext() ? seroIter.next() : seroType1_CIWD;
-
-      typeIter = typesSetNonCWD.iterator();
-      HLAType hlaType1_First = typeIter.next();
-      HLAType hlaType2_First = typeIter.hasNext() ? typeIter.next() : hlaType1_CIWD;
-
-      seroIter = locusSetNonCWD.iterator();
-      SeroType seroType1_First = seroIter.next();
-      SeroType seroType2_First = seroIter.hasNext() ? seroIter.next() : seroType1_First;
-
-      // TODO encapsulate Allele selection JavaFX components in a different class in the JFX
-      // packages
-      // ValidationModelBuilder shouldn't know about JavaFX UI components
-
-      HBox assignedPane = new HBox(10);
-      assignedPane.setMaxWidth(Double.MAX_VALUE);
-      assignedPane.setAlignment(Pos.CENTER_LEFT);
-      Label assignedHeader = new Label("Assigned allele pair: ");
-      TextFlow assignedTextFlow = new TextFlow();
-      addTextNodes(assignedTextFlow, hlaType1_First.toString(), true);
-      assignedTextFlow.getChildren().add(new Text(" | "));
-      addTextNodes(assignedTextFlow, hlaType2_First.toString(), true);
-      Label assignedHeaderLbl = new Label("", assignedTextFlow);
-      assignedHeaderLbl.setMaxWidth(Double.MAX_VALUE);
-      assignedHeaderLbl.setMaxHeight(Double.MAX_VALUE);
-      assignedPane.getChildren().add(assignedHeader);
-      assignedPane.getChildren().add(getSpacer());
-      assignedPane.getChildren().add(assignedHeaderLbl);
-
-      HBox ciwdPane = new HBox(10);
-      ciwdPane.setMaxWidth(Double.MAX_VALUE);
-      ciwdPane.setAlignment(Pos.CENTER_LEFT);
-      Label ciwdHeader = new Label("Common / Well-Documented allele pair: ");
-      TextFlow ciwdTextFlow = new TextFlow();
-      addTextNodes(ciwdTextFlow, hlaType1_CIWD.toString(), true);
-      ciwdTextFlow.getChildren().add(new Text(" | "));
-      addTextNodes(ciwdTextFlow, hlaType2_CIWD.toString(), true);
-      Label ciwdHeaderLbl = new Label("", ciwdTextFlow);
-      ciwdHeaderLbl.setMaxWidth(Double.MAX_VALUE);
-      ciwdHeaderLbl.setMaxHeight(Double.MAX_VALUE);
-      ciwdPane.getChildren().add(ciwdHeader);
-      ciwdPane.getChildren().add(getSpacer());
-      ciwdPane.getChildren().add(ciwdHeaderLbl);
-
-      String text = "Or manually select allele pair for this locus";
-      String text1 =
-          "(selecting the first allele will populate valid pairings for the second allele)";
-
-      final List<Supplier<TextFlow>> allChoices = choices.getAllChoices();
-
-      Function<Supplier<TextFlow>, String> tooltipProvider = (tf) -> {
-        return Objects.toString(choices.dataMap.get(tf), null);
-      };
-
-      String filterName = "Show CIWD Alleles Only";
-      Predicate<Supplier<TextFlow>> filter = (tf) -> {
-        return isAnyPartCWD(choices.getData(tf));
-      };
-
-      StyleableContingentChoiceDialog.Option<Supplier<TextFlow>> opt1 =
-          new Option<>(assignedPane, choices.getChoiceForAllele(hlaType1_First.toString()),
-              choices.getChoiceForAllele(hlaType2_First.toString()));
-
-      StyleableContingentChoiceDialog.Option<Supplier<TextFlow>> opt2 =
-          new Option<>(ciwdPane, choices.getChoiceForAllele(hlaType1_CIWD.toString()),
-              choices.getChoiceForAllele(hlaType2_CIWD.toString()));
-
-      Label manualChoice = new Label(text + System.lineSeparator() + text1);
-
-      StyleableContingentChoiceDialog<Supplier<TextFlow>> cd =
-          new StyleableContingentChoiceDialog<>(null, allChoices, Lists.newArrayList(),
-              choices.getAllSecondChoices(), choices.dataMap, filterName,
-              filter/* , textEntryMatcher */, opt1, opt2, manualChoice);
-      cd.setTitle("Select HLA-" + locus.name() + " Alleles");
-
-      cd.setHeaderText("Assigned allele pair for HLA-" + locus.name()
-          + " locus is not Common / Well-Documented.");
-      cd.setContentText("Please select desired allele pair for this locus:");
-
-      cd.setCombo1CellFactory(listView -> new SimpleTableObjectListCell(tooltipProvider));
-      cd.setCombo1ButtonCell(new SimpleTableObjectListCell(tooltipProvider));
-      cd.setCombo2CellFactory(listView -> new SimpleTableObjectListCell(tooltipProvider));
-      cd.setCombo2ButtonCell(new SimpleTableObjectListCell(tooltipProvider));
-      cd.setConverter1(new AlleleStringConverter(choices));
-      cd.setConverter2(new AlleleStringConverter(choices));
-
-      Optional<Supplier<TextFlow>> result = cd.showAndWait();
-
-      if (!result.isPresent()) {
-        throw new CancellationException();
-      }
-
-      Supplier<TextFlow> selAllele1 = result.get();
-      Supplier<TextFlow> selAllele2 = cd.getSelectedSecondItem();
-
-      String allele1 = choices.getData(selAllele1);
-      String allele2 = choices.getData(selAllele2);
-
-      HLAType hlaType1 = HLAType.valueOf(allele1);
-      HLAType hlaType2 = HLAType.valueOf(allele2);
-      SeroType seroType1 = hlaType1.equivSafe();
-      SeroType seroType2 = hlaType2.equivSafe();
-
-      locusSet.clear();
-      locusSet.add(seroType1);
-      locusSet.add(seroType2);
-      typesSet.clear();
-      typesSet.add(hlaType1);
-      typesSet.add(hlaType2);
-
-      boolean a1Match =
-          hlaType1.compareTo(hlaType1_First) == 0 || hlaType1.compareTo(hlaType2_First) == 0;
-      boolean a2Match =
-          hlaType2.compareTo(hlaType1_First) == 0 || hlaType2.compareTo(hlaType2_First) == 0;
-
-      if (!a1Match || !a2Match) {
-        ImmutableSortedSet<TypePair> prevSet =
-            ImmutableSortedSet.of(new TypePair(hlaType1_First, seroType1_First),
-                new TypePair(hlaType2_First, seroType2_First));
-
-        ImmutableSortedSet<TypePair> newSet = ImmutableSortedSet
-            .of(new TypePair(hlaType1, seroType1), new TypePair(hlaType2, seroType2));
-
-
-        return Pair.of(prevSet, newSet);
-      }
-
-      return null;
-    }
-  }
-
-  private Region getSpacer() {
-    Region spacer = new Region();
-    HBox.setHgrow(spacer, Priority.ALWAYS);
-    return spacer;
-  }
-
-  private AllelePairings getAllelePairs(HLALocus locus) {
+  public AllelePairings getAllelePairs(HLALocus locus) {
     return possibleAllelePairings.get(locus);
-  }
-
-  public abstract static class PresentableDataChoices<T, R> {
-    List<T> choices;
-    Multimap<T, T> secondChoices;
-    BiMap<T, R> dataMap;
-
-    public PresentableDataChoices(List<T> choices, Multimap<T, T> secondChoices,
-        BiMap<T, R> presentationToDataMap) {
-      this.choices = choices;
-      this.secondChoices = secondChoices;
-      this.dataMap = presentationToDataMap;
-    }
-
-    public List<T> getAllChoices() {
-      return choices;
-    }
-
-    public Multimap<T, T> getAllSecondChoices() {
-      return secondChoices;
-    }
-
-    public Collection<T> getSecondChoices(T firstChoice) {
-      return secondChoices.get(firstChoice);
-    }
-
-    public abstract List<T> getMatchingChoices(String userInput);
-
-    public R getData(T presentable) {
-      return dataMap.get(presentable);
-    }
-
-    public T getPresentation(R data) {
-      return dataMap.inverse().get(data);
-    }
-
-  }
-
-  public static class PresentableAlleleChoices
-      extends PresentableDataChoices<Supplier<TextFlow>, String> {
-
-    public static PresentableAlleleChoices create(HLALocus locus, AllelePairings allelePairings) {
-
-      List<Supplier<TextFlow>> userChoices = new ArrayList<>();
-      ListMultimap<Supplier<TextFlow>, Supplier<TextFlow>> secondChoices =
-          SortedSetMultimapBuilder.hashKeys().arrayListValues().build();
-      BiMap<Supplier<TextFlow>, String> presentationToDataMap = HashBiMap.create();
-
-      List<String> alleleKeys = new ArrayList<>(allelePairings.map.keySet());
-      sortAlleleStrings(alleleKeys);
-
-      LinkedListMultimap<String, String> choiceList =
-          condenseIntoGroups(allelePairings, alleleKeys, true);
-
-      for (String allele : choiceList.keySet()) {
-        Supplier<TextFlow> supp = () -> getText(allele);
-
-        userChoices.add(supp);
-        presentationToDataMap.put(supp, allele);
-      }
-
-
-      for (Supplier<TextFlow> choice : userChoices) {
-        String data = presentationToDataMap.get(choice);
-        if (data == null) {
-          continue;
-        }
-        String pairingAllele = data.contains("-") ? data.split("-")[0] : data;
-
-        List<String> pairings = new ArrayList<>(allelePairings.getValidPairings(pairingAllele));
-        sortAlleleStrings(pairings);
-
-        LinkedListMultimap<String, String> secondChoicePairings =
-            condenseIntoGroups(allelePairings, pairings, false);
-
-        for (String pairing : secondChoicePairings.keySet()) {
-          Supplier<TextFlow> presentationView = presentationToDataMap.inverse().get(pairing);
-          if (presentationView == null) {
-            presentationView = () -> getText(pairing);
-            presentationToDataMap.put(presentationView, pairing);
-          }
-          secondChoices.put(choice, presentationView);
-        }
-      }
-
-      Map<String, String> alleleToGroupMap = new HashMap<>();
-      choiceList.entries().stream().forEach(e -> {
-        alleleToGroupMap.put(e.getValue(), e.getKey());
-      });
-
-      return new PresentableAlleleChoices(userChoices, secondChoices, presentationToDataMap,
-          allelePairings, alleleToGroupMap);
-    }
-
-    private static LinkedListMultimap<String, String> condenseIntoGroups(
-        AllelePairings allelePairings, List<String> alleleKeys, boolean checkPairings) {
-
-      List<List<String>> subsets = new ArrayList<>();
-      List<String> subset = new ArrayList<>();
-
-      HLAType prev3FieldType = null;
-
-      // all allele strings should be in the correct order now
-      // now let's condense into subsets of the same serotypes, with
-      // separate subsets for N/n and LSCAQ/lscaq alleles
-      for (String allele : alleleKeys) {
-        if (allele.matches(NOT_EXPRESSED) || allele.matches(NOT_ON_CELL_SURFACE)) {
-          // make sure to add the currently-being-built subset first!
-          if (subset.size() > 0) {
-            subsets.add(subset);
-            subset = new ArrayList<>();
-          }
-          // add null or lscaq alleles as single entries
-          subsets.add(Lists.newArrayList(allele));
-          // clear out known three-field type
-          prev3FieldType = null;
-          continue;
-        }
-
-        HLAType hType = HLAType.valueOf(allele);
-        if (hType.resolution() <= 2) {
-          // make sure to add the currently-being-built subset first!
-          if (subset.size() > 0) {
-            subsets.add(subset);
-            subset = new ArrayList<>();
-          }
-          // add two-field alleles as single entries
-          subsets.add(Lists.newArrayList(allele));
-          // clear out known three-field type
-          prev3FieldType = null;
-          continue;
-        }
-
-        // we know resolution is >= 3
-
-        HLAType curr3FieldType = (hType.resolution() == 3) ? hType
-            : new HLAType(hType.locus(), hType.spec().get(0), hType.spec().get(1),
-                hType.spec().get(2));
-        if (prev3FieldType != null && prev3FieldType.compareTo(curr3FieldType) != 0) {
-          // new three field type
-          subsets.add(subset);
-          subset = new ArrayList<>();
-        }
-        // same or new three-field type, either way:
-        // update known three-field type, and
-        // add four-field type to subset
-        prev3FieldType = curr3FieldType;
-        subset.add(allele);
-
-      }
-      if (subset.size() > 0) {
-        subsets.add(subset);
-      }
-
-      LinkedListMultimap<String, String> groupToAllelesMap = LinkedListMultimap.create();
-
-      // now we need to process the subset lists
-      // single element subsets can be added directly
-      // -- this preserves null/not-expressed subsets
-      // multiple element subsets need to be further
-      // separated into subsets: all alleles in a subset must
-      // map to the same serotype and the same second choice alleles
-
-      for (List<String> sub : subsets) {
-        if (sub.size() == 1) {
-          groupToAllelesMap.put(sub.get(0), sub.get(0));
-          continue;
-        }
-
-        List<List<String>> subSubsets = new ArrayList<>();
-        List<String> subSubset = new ArrayList<>();
-
-        SeroType prevSero = null;
-        Set<String> prevPairings = null;
-
-        for (String subAllele : sub) {
-          HLAType type = HLAType.valueOf(subAllele);
-          SeroType newSero = type.equivSafe();
-          HashSet<String> newPairings =
-              checkPairings ? Sets.newHashSet(allelePairings.getValidPairings(subAllele))
-                  : new HashSet<>();
-          if (prevSero == null) {
-            prevSero = newSero;
-            prevPairings = newPairings;
-          } else if (prevSero.compareTo(newSero) != 0) {
-            if (subSubset.size() > 0) {
-              subSubsets.add(subSubset);
-              subSubset = new ArrayList<>();
-            }
-            prevSero = newSero;
-            prevPairings = newPairings;
-          } else if (!prevPairings.containsAll(newPairings)
-              || !newPairings.containsAll(prevPairings)) {
-            if (subSubset.size() > 0) {
-              subSubsets.add(subSubset);
-              subSubset = new ArrayList<>();
-            }
-            prevSero = newSero;
-            prevPairings = newPairings;
-          }
-          subSubset.add(subAllele);
-        }
-        if (subSubset.size() > 0) {
-          subSubsets.add(subSubset);
-          subSubset = new ArrayList<>();
-        }
-
-        for (List<String> subS : subSubsets) {
-          if (subS.size() == 1) {
-            groupToAllelesMap.put(subS.get(0), subS.get(0));
-          } else {
-            String groupName = subS.get(0) + "-" + subS.get(subS.size() - 1);
-            for (String s : subS) {
-              groupToAllelesMap.put(groupName, s);
-            }
-          }
-        }
-      }
-
-      return groupToAllelesMap;
-    }
-
-    private static void sortAlleleStrings(List<String> alleleKeys) {
-      alleleKeys.sort((s1, s2) -> {
-        boolean check1N = s1.matches(NOT_EXPRESSED);
-        boolean check1C = s1.matches(NOT_ON_CELL_SURFACE);
-        boolean check2N = s2.matches(NOT_EXPRESSED);
-        boolean check2C = s2.matches(NOT_ON_CELL_SURFACE);
-        boolean check1 = check1N || check1C;
-        boolean check2 = check2N || check2C;
-        // char s1C = s1.charAt(s1.length() - 1);
-        // char s2C = s2.charAt(s2.length() - 1);
-        String s1Hs = check1 ? s1.substring(0, s1.length() - 1) : s1;
-        String s2Hs = check2 ? s2.substring(0, s2.length() - 1) : s2;
-
-        HLAType h1 = HLAType.valueOf(s1Hs);
-        HLAType h2 = HLAType.valueOf(s2Hs);
-
-        int comp;
-        if ((comp = h1.compareTo(h2)) != 0)
-          return comp;
-
-        if (check1 && !check2) {
-          // first element ends with special character, meaning second element should come first
-          return 1;
-        } else if (check2 && !check1) {
-          // second element ends with special character, meaning first element should come first
-          return -1;
-        }
-
-        // both end with a special character
-        if (check1N && check2N) {
-          // both null and same HLAType - these are the same allele
-          return 0;
-        } else if (check1N && !check2N) {
-          // first element is null, second is lscaq, second comes first
-          return 1;
-        } else if (!check1N && check2N) {
-          // first element is lscaq, second is null, first comes first
-          return -1;
-        }
-
-
-        // this block could probably be condensed based on knowing null status from previous checks,
-        // but it's easier to just duplicate the logic for now...
-
-        // both end with a special character
-        if (check1C && check2C) {
-          // both null and same HLAType - these are the same allele
-          return 0;
-        } else if (check1C && !check2C) {
-          // first element is null, second is lscaq, second comes first
-          return 1;
-        } else if (!check1C && check2C) {
-          // first element is lscaq, second is null, first comes first
-          return -1;
-        }
-
-        // TODO dunno what's going on here... probably should do something special?
-        return 0;
-      });
-    }
-
-    private final AllelePairings allelePairs;
-
-    Map<String, String> alleleToGroupKeyMap;
-
-    public Supplier<TextFlow> getChoiceForAllele(String allele) {
-      return getPresentation(alleleToGroupKeyMap.get(allele));
-    }
-
-    private PresentableAlleleChoices(List<Supplier<TextFlow>> choices,
-        Multimap<Supplier<TextFlow>, Supplier<TextFlow>> secondChoices,
-        BiMap<Supplier<TextFlow>, String> presentationToDataMap, AllelePairings allelePairs,
-        Map<String, String> alleleToGroupMap) {
-      super(choices, secondChoices, presentationToDataMap);
-      this.allelePairs = allelePairs;
-      this.alleleToGroupKeyMap = alleleToGroupMap;
-    }
-
-    @Override
-    public List<Supplier<TextFlow>> getMatchingChoices(String userInput) {
-      return allelePairs.getMatchingAlleles(userInput).stream().map(s -> dataMap.inverse().get(s))
-          .filter(Predicates.notNull()).collect(Collectors.toList());
-    }
-
-  }
-
-  private static TextFlow getText(String allele) {
-    TextFlow flow = new TextFlow();
-    if (allele.contains("-")) {
-      String[] a = allele.split("-");
-      addTextNodes(flow, a[0], false);
-      flow.getChildren().add(new Text(" - "));
-      addTextNodes(flow, a[1], true);
-    } else {
-      addTextNodes(flow, allele, true);
-    }
-
-    return flow;
-  }
-
-  private boolean isAnyPartCWD(String allele) {
-    HLAType alleleType = HLAType.valueOf(allele);
-    Status status1 = CommonWellDocumented.getStatus(alleleType);
-
-    return status1 != Status.UNKNOWN;
-  }
-
-  private static void addTextNodes(TextFlow flow, final String allele, boolean addSero) {
-    HLAType alleleType = HLAType.valueOf(allele);
-    HLAType cwdType1 = CommonWellDocumented.getCWDType(alleleType);
-    Status status1 = CommonWellDocumented.getStatus(alleleType);
-
-    flow.getChildren().add(new Text(alleleType.locus().name() + "*"));
-
-    String specString = alleleType.specString();
-    boolean match = allele.matches(NOT_EXPRESSED) || allele.matches(NOT_ON_CELL_SURFACE);
-
-    if (status1 != Status.UNKNOWN) {
-
-      if (cwdType1.specString().length() < specString.length()) {
-        Text t1 = new Text(cwdType1.specString());
-        t1.setStyle("-fx-font-weight:bold;");
-        flow.getChildren().add(t1);
-        flow.getChildren().add(new Text(specString.substring(cwdType1.specString().length())));
-      } else {
-        Text t1 = new Text(specString);
-        t1.setStyle("-fx-font-weight:bold;");
-        flow.getChildren().add(t1);
-      }
-
-      if (match) {
-        flow.getChildren().add(new Text("" + allele.charAt(allele.length() - 1)));
-      }
-
-    } else {
-      flow.getChildren()
-          .add(new Text(specString + (match ? ("" + allele.charAt(allele.length() - 1)) : "")));
-    }
-
-    if (addSero) {
-      flow.getChildren().add(
-          new Text(" (" + alleleType.locus().name() + alleleType.equivSafe().specString() + ")"));
-    }
-  }
-
-  private static class SimpleTableObjectListCell extends ListCell<Supplier<TextFlow>> {
-
-    // private Function<Supplier<TextFlow>, Supplier<TextFlow>> tooltipProvider;
-    private Function<Supplier<TextFlow>, String> tooltipProvider;
-    private Tooltip tooltip = new Tooltip();
-
-    public SimpleTableObjectListCell(Function<Supplier<TextFlow>, String> tooltipProvider) {
-      // Function<Supplier<TextFlow>, Supplier<TextFlow>> tooltipProvider) {
-      this.tooltipProvider = tooltipProvider;
-
-      // hack for adjusting tooltip delay / etc
-      // from https://stackoverflow.com/a/43291239/875496
-      // TODO FIXME change when JavaFX9+ is available
-      try {
-        Class<?> clazz = tooltip.getClass().getDeclaredClasses()[0];
-        Constructor<?> constructor = clazz.getDeclaredConstructor(Duration.class, Duration.class,
-            Duration.class, boolean.class);
-        constructor.setAccessible(true);
-        Object tooltipBehavior = constructor.newInstance(new Duration(50), // open
-            new Duration(500000), // visible
-            new Duration(100), // close
-            false);
-        Field fieldBehavior = tooltip.getClass().getDeclaredField("BEHAVIOR");
-        fieldBehavior.setAccessible(true);
-        fieldBehavior.set(tooltip, tooltipBehavior);
-      } catch (Throwable t) {
-        t.printStackTrace();
-      }
-
-      tooltip.setWrapText(true);
-      tooltip.setMaxWidth(600);
-
-      /*
-       * NB: TODO FIXME setting style here because I couldn't figure out how to locate it in an
-       * actual CSS style sheet -- 04-18-2024, b.cole
-       */
-      tooltip.setStyle(
-          "-fx-background: rgba(230,230,230); -fx-text-fill: black; -fx-background-color: rgba(230,230,230,0.95); -fx-background-radius: 5px; -fx-background-insets: 0; -fx-padding: 0.667em 0.75em 0.667em 0.75em; -fx-effect: dropshadow( three-pass-box , rgba(0,0,0,0.5) , 10, 0.0 , 0 , 3 ); -fx-font-size: 0.95em;");
-    }
-
-    @Override
-    public void updateItem(Supplier<TextFlow> item, boolean empty) {
-      try {
-        super.updateItem(item, empty);
-        if (item != null && !empty) {
-          setGraphic(item.get());
-
-          String tf = tooltipProvider.apply(item);
-          // Supplier<TextFlow> tf = tooltipProvider.apply(item);
-          if (tf == null || tf.length() < 150) {
-            // TODO don't show tooltips if item isn't too long!
-            setTooltip(null);
-          } else {
-            // System.out.println("tool:: " + tf);
-            tooltip.setText(tf);
-            // final TextFlow value = tf.get();
-            // value.setMaxWidth(tooltip.getMaxWidth());
-            // tooltip.setGraphic(value);
-            setTooltip(tooltip);
-          }
-        } else {
-          setGraphic(null);
-          setTooltip(null);
-        }
-      } catch (Throwable t) {
-        t.printStackTrace();
-      }
-    }
-
   }
 
   /** Helper method to build a set if it's null */
@@ -1781,6 +1174,10 @@ public class ValidationModelBuilder {
     public void addPairing(String a1, String a2) {
       map.put(a1, a2);
       map.put(a2, a1);
+    }
+
+    public Set<String> getAlleleKeys() {
+      return map.keySet();
     }
 
     public Collection<String> getValidPairings(String allele) {
@@ -1975,5 +1372,9 @@ public class ValidationModelBuilder {
       Collections.sort(sorted);
       return sorted;
     }
+  }
+
+  public Set<HLALocus> getNonCWDLoci() {
+    return nonCWDLoci;
   }
 }
