+++ src/main/java/org/pankratzlab/unet/model/remap/CancellationException.java
@@ -1,0 +1,11 @@
+package org.pankratzlab.unet.model.remap;
+
+/**
+ * Used to indicate user-cancellation of locus remapping
+ */
+public class CancellationException extends RuntimeException {
+
+  private static final long serialVersionUID = 1L;
+
+
+}
